# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "hotel-reservation-api",
    "description": "API for managing hotel reservations"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/reservations": {
      "get": {
        "summary": "Get all reservations or search reservations",
        "parameters": [
          {
            "name": "nameHotel",
            "in": "query",
            "type": "string",
            "description": "Filter reservations by hotel name"
          },
          {
            "name": "arrivalDate",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "Filter reservations by arrival date"
          },
          {
            "name": "departureDate",
            "in": "query",
            "type": "string",
            "format": "date",
            "description": "Filter reservations by departure date"
          },
          {
            "name": "typeRoom",
            "in": "query",
            "type": "string",
            "description": "Filter reservations by room type"
          },
          {
            "name": "paymentStatus",
            "in": "query",
            "type": "string",
            "description": "Filter reservations by payment status"
          },
          {
            "name": "passengers",
            "in": "query",
            "type": "integer",
            "description": "Filter reservations by number of passengers"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reservations"
          }
        }
      },
      "post": {
        "summary": "Create a new reservation",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Reservation object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Reservation"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Reservation created"
          }
        }
      }
    },
    "/reservations/{id}": {
      "get": {
        "summary": "Get a reservation by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation details"
          },
          "404": {
            "description": "Reservation not found"
          }
        }
      },
      "put": {
        "summary": "Update a reservation by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Reservation object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Reservation"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation updated"
          },
          "404": {
            "description": "Reservation not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a reservation by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Reservation deleted"
          },
          "404": {
            "description": "Reservation not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Reservation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "arrivalDate": {
          "type": "string",
          "format": "date"
        },
        "departureDate": {
          "type": "string",
          "format": "date"
        },
        "nameHotel": {
          "type": "string"
        },
        "typeRoom": {
          "type": "string"
        },
        "passengers": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "mail": {
          "type": "string",
          "format": "email"
        },
        "paymentStatus": {
          "type": "string"
        }
      }
    }
  }
}